+++ CHANGELOG.md
@@ -77,6 +77,8 @@
     - Makes sure that candidates are deduplicated upon discovery.
     - Added a 3rd phase: Initial candidates which pass verification are verified again.
 - Modules
+  - Added
+    - Auto-complete for password form fields.
   - General
       - Updated module names along with some descriptions and issue names.
       - Limited the maximum number of issues to 25 for the following recon modules: